seperate sidebar folder
