+++ src/components/layout/BaseLayout.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react"
 import Toast, { Position } from "../toast/ToastList"
 import { Navbar } from "../navigation/Navbar"
-import { Sidebar } from "../navigation/Sidebar"
+import { Sidebar } from "../sidebar/Sidebar"
 import useToastStore from "../toast/ToastStore"
 
 interface Props {
