+++ src/pages/Page/Page.tsx
@@ -58,7 +58,7 @@ export const Page = () => {
             }
 
             const currentPage = res[0]
-            setCurrentPage(currentPage)
+            // setCurrentPage(currentPage)
             setIsLoading(false)
         })
     }
@@ -92,7 +92,7 @@ export const Page = () => {
 
     return (
         <div className="w-full h-full max-h-full overflow-y-scroll p-10 flex justify-center relative bg-slate-50">
-            {isArchived && <Banner />}
+            {/* {isArchived && <Banner />}
             <div className="h-full w-[50rem] bg-white rounded-md">
                 {!isLoading && currentPage ? (
                     <div className="min-h-full w-full flex flex-col shadow-sm p-2">
@@ -120,7 +120,7 @@ export const Page = () => {
                 ) : (
                     <p>Loading...</p>
                 )}
-            </div>
+            </div> */}
         </div>
     )
 }
